// src/app/models/Client.ts
export interface Client {
    key?:string; // ? --> optional
    firstName?:string;
    lastName?:string;
    aka?:string;
    email?: string;
    age?: string;
    gender?:string;
    adharCard?: string;
    phone?: string;
    balance?: number;
    dateOfBirth?: string;
    maritalStatus?: string;
    designation?: number;
  }